requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.3.0
azure-ai-inference==1.0.0b9
azure-core==1.37.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.1
croniter==6.0.0
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.128.0
filelock==3.24.3
filetype==1.2.0
frozenlist==1.7.0
fsspec==2026.2.0
google-auth==2.47.0
google-genai==1.60.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.11
isodate==0.7.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.14
langchain-google-genai==4.2.1
langchain-huggingface==1.2.0
langchain-openai==1.1.9
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.1
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.6
langsmith==0.4.32
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
openai==2.14.0
orjson==3.11.3
ormsgpack==1.12.2
packaging==25.0
portalocker==3.2.0
propcache==0.4.0
protobuf==6.33.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pypdf==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.2
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rq==2.6.1
rsa==4.9.1
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.40.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.20.1
zstandard==0.25.0
